+++ server/src/prisma/repositories/trash/prisma-delete-trash-repository.ts
@@ -2,14 +2,27 @@ import { prisma } from "@src/prisma/prisma-client";
 import { DeleteRepository } from "@src/repositories/delete-repository";
 
 export class PrismaDeleteTrashRepository implements DeleteRepository {
-  async delete(id: string): Promise<void> {
-    await prisma.deletedTask.delete({
+  async delete(userId: string, id: string): Promise<void> {
+    const trash = await prisma.deletedTask.findFirst({
       where: {
-        id: id,
+        userId,
       },
     });
+
+    if (trash) {
+      await prisma.deletedTask.delete({
+        where: {
+          id,
+        },
+      });
+    }
   }
-  async deleteAll(): Promise<void> {
-    await prisma.deletedTask.deleteMany();
+
+  async deleteAll(userId: string): Promise<void> {
+    await prisma.deletedTask.deleteMany({
+      where: {
+        userId,
+      },
+    });
   }
 }
